renderShadows 95.14->95.29 (u16 cast on 2nd getAngle -> clrlwi before -0x3fc8)

+++ src/main/dll/dll_0031_minimap.c
@@ -449,7 +449,7 @@ int Minimap_update(void)
                     }
                     uq = ox / gMinimapZoom;
                     u = uq;
-                    frac = gMinimapZoom * (uq - u);
+                    frac = gMinimapZoom * (uq - (f32)u);
                     vq = oy / gMinimapZoom;
                     vv = vq;
                     ((u8*)&col)[3] = gMinimapContentAlpha;
